Improve stability of xorgxrdp when resolution changes during a session ("dynamic resolution")
During dynamic resize, it's possible for the capture timer to trigger while the screen's memory buffer is not in a proper state for the given codec used.

This patches the issue so that the xorgxrdp driver won't crash, but I could envision further enhancements to the _rdpClientCon data structure that make this easier to access.